% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramnet_mbo.R
\name{mbo_recommended}
\alias{mbo_recommended}
\title{Fit recommended regularized tram based on model based optimization output}
\usage{
mbo_recommended(mbo_obj, m0, x, ...)
}
\arguments{
\item{mbo_obj}{Object returned by \code{\link[tramnet]{mbo_tramnet}}.}

\item{m0}{Null model of class \code{"tram"}, see
\code{\link[tramnet]{tramnet.tram}}.}

\item{x}{Design matrix, see \code{\link[tramnet]{tramnet.tram}}.}

\item{...}{Additional arguments to \code{tramnet}.}
}
\value{
Object of class \code{"tramnet"}.
}
\description{
Extracts the "optimal" tuning parameters from the output of
\code{mbo_tramnet} and fits the corresponding \code{tramnet} model.
}
