% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tramnet}
\alias{plot.tramnet}
\title{Plot \code{"tramnet"} objects}
\usage{
\method{plot}{tramnet}(
  x,
  newdata = NULL,
  type = c("distribution", "survivor", "density", "logdensity", "hazard", "loghazard",
    "cumhazard", "quantile", "trafo"),
  q = NULL,
  prob = 1:(K - 1)/K,
  K = 50,
  col = rgb(0.1, 0.1, 0.1, 0.1),
  lty = 1,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"tramnet"}.}

\item{newdata}{See \code{\link[mlt]{plot.ctm}}.}

\item{type}{See \code{\link[mlt]{plot.ctm}}.}

\item{q}{See \code{\link[mlt]{plot.ctm}}.}

\item{prob}{See \code{\link[mlt]{plot.ctm}}.}

\item{K}{See \code{\link[mlt]{plot.ctm}}.}

\item{col}{See \code{\link[mlt]{plot.ctm}}.}

\item{lty}{See \code{\link[mlt]{plot.ctm}}.}

\item{add}{See \code{\link[mlt]{plot.ctm}}.}

\item{...}{Additional arguments passed to \code{\link[mlt]{plot.ctm}}.}
}
\description{
Plot \code{"tramnet"} objects
}
