
\name{print.summary.tramnet}
\alias{print.summary.tramnet}
\title{print summary method for class \code{"tramnet"}}
\usage{
\method{print}{summary.tramnet}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{object of class \code{"tramnet"}}

\item{digits}{number of digits to print}

\item{...}{additional arguments}
}
\description{
print summary method for class \code{"tramnet"}
}
\author{
Lucas Kook
}
\value{
prints textual summary in the console and returns an invisible copy of the 
\code{"tramnet"} object
}