
\name{print.tramnet}
\alias{print.tramnet}
\title{print method for class \code{"tramnet"}}
\usage{
\method{print}{tramnet}(x, ...)
}
\arguments{
\item{x}{object of class \code{"tramnet"}}

\item{...}{additional arguments to \code{\link{summary}}}
}
\description{
print method for class \code{"tramnet"}
}
\author{
Lucas Kook
}
