
\name{summary.tramnet}
\alias{summary.tramnet}
\title{summary method for class \code{"tramnet"}}
\usage{
\method{summary}{tramnet}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramnet"}}

\item{...}{additional arguments}
}
\description{
summary method for class \code{"tramnet"}
}
\author{
Lucas Kook
}
\value{
Returns an object of class \code{"summary.tramnet"} containing
information about the model, optimization status, sparsity and
tuning parameters
}