
\name{mbo_recommended}
\alias{mbo_recommended}
\title{Fit recommended regularized tram based on model based optimization output}
\usage{
mbo_recommended(mbo_obj, m0, x, ...)
}
\arguments{
\item{mbo_obj}{object return by \code{mbo_tramnet}}

\item{m0}{null model of class \code{"tram"}}

\item{x}{matrix of covariables}

\item{...}{additional arguments to \code{tramnet()}}
}
\description{
Fit recommended regularized tram based on model based optimization output
}
